cleanup: Moved middelwares
